requirements.txt
fastapi
uvicorn[standard]
redis>=4.6.0
prometheus-client
httpx
python-dotenv
